Minor overhaul of the elaboration of FFI related constructs.
Eliminated the distinction between Prim.FFI_Symbol {fetch = true, ...}
and Prim.FFI_Symbol {fetch = false, ...} in favor of the latter.  This
means that Prim.FFI_symbol {name, ...} is interpreted as a primitive
that evaluates to the address of the C-symbol denoted by name.

Made
  _import "name" : t ;
of a non-functional type elaborate to
  (Prim.Pointer_get*(Prim.FFI_Symbol {name},0)) : t
where * is the C-type of t.

The one minor complication is that _import of a bool must coerced from
a word to a CoreML boolean.