Fixed a bug in the representation of flat arrays/vectors that caused incorrect behavior when the element size was 2 or 4 bytes and there were multiple components to the element (e.g. (char * char) vector).
The bug was in the representation pass, in the part that computes byte
offsets when selecting fields out of packed records.  The fix took out
a patch that I had put in to make things work on big-endian machines
(Sparc), but I think I got it right even for those, although I haven't
tested on them.