Fix(?) for the subtle bug in the refFlatten pass.  The pass was sensitive to the ordering of functions in the SSA2 program representation.  While it would be o.k. (though, less than ideal) to get conservative results depending on the order of functions (i.e., flatten less), the bug manifests itself as an unsoundness in the optimization.
I believe that the bug is related to the processing of arguments in
the foreachObject loop:

             (* Check that all arguments that are represented by flattening them
              * into the object are available as an explicit allocation.
              *)
             val () =
                Vector.foreach
                (args, fn a =>
                 case Value.deFlat {inner = varValue a, outer = obj} of
                    NONE => ()
                  | SOME (Obj {flat, ...}) =>
                       case ! (varInfo a) of
                          Flattenable _ => ()
                        | Unflattenable => flat := NotFlat)

Value.deFlat returns the inner object, and only succeeds if the inner
object's flat field is an Offset of the outer object.  If we process
an argument whose object's flat field is Unknown, then we will not
promote the field to NotFlat.  If we later process the argument's
object as an object in its own right, we may at that later point in
time promote the flat field to Offset.  But, because the object is
used as an argument, we really should promote it all the way to
NotFlat -- as would have happened if we processed the object as an
object first and then as an argument.

The fix is to apply two foreachObject loops.  The first one processes
the useStatus, while the second processes arguments.  This preserves
the intended behavior, as processing an argument can only promote a
flat field to NotFlat, which does not affect the processing of the
useStatus.  Meanwhile, processing useStatus first promotes all
not-Unknown flat fields before beginning to process arguments.