Made  SsaToRssa.convert  respect  codegenImplementsPrim  on primitives used in Arith transfers.  An Arith primitive that is not implemented by a codegen is deconstructed into the non-checking primitive (which is subsequently checked for codegen implementation) and a CCall which checks whether or not the operation overflowed.  This is not particularly efficient: an operation that isn't implemented either with or without overflow checking requires two separate CCalls. There are better implmentations: one CCall that returns the operation and sets a boolean overflow flag, etc.  However, no current codegen fails to implement an Arith primitive.  The decomposition of Arith transfers is provided as a crux for future codegens that could avoid implementing Arith transfers during intial development.
Note: the current codegens could use this facility to avoid
implementing 64bit multiplication with overflow.  We currently hack
the Basis Library implementation to perform 64bit multiplication with
overflow by casting to and from IntInf.ints.  I did a simple 64bit
factorial benchmark and determined that the IntInf method and the two
CCall methods were just about even.  But, a better deconstruction of
Arith transfers might make that a better implementation.