MAIL improved exception history for Overflow
Put in a hack suggested by Matthew to cause Overflow exceptions to
have their exception history printed just like all other exceptions.
The problem had been that because Overflow is generated in the
compiler internals (during closure conversion's translation to SSA),
it misses the pass (implementExceptions) that inserts the calls to the
hooks (defined in basis-library/mlton/exn.sml) that attach the
exception history to an exception when it is first raised.

So, a program like the following

----------------------------------------------------------------------
fun f n =
   case n of
      0 => 0
    | 1 => raise Fail ""
    | n => 1 + f (n + 1)
val _ = f (valOf Int.maxInt - 5)
----------------------------------------------------------------------

when compiled with MLton 20041109 and -const 'Exn.keepHistory true',
will print only

unhandled exception: Overflow

When compiled by the CVS HEAD before this commit, the program prints
the same.

After this commit, the program will print

unhandled exception: Overflow
with history:
        f z.sml 1.5
        f z.sml 1.5
        f z.sml 1.5
        f z.sml 1.5
        f z.sml 1.5
        f z.sml 1.5

The fix was only a few lines changed in
basis-library/misc/primitive.sml.  The idea is to treat the Overflow
exception exported by the compiler as "PrimitiveOverflow", declare the
real Overflow exception in the basis sources, and then wrap every
primitive that can raise overflow with a handler that re-raises with
the basis Overflow.  It is that reraise that will cause the exception
history to be exported.