Eliminated expert flag  -profile mark. Added expert flags  -profile {drop,label}.
The latter provides the functionality of -profile mark; i.e., it
inserts profiling labels into the final executable, but disables time
profiling.

The former causes the implementProfiling pass to implement nothing.
It should be an improvement over -profile mark for measuring the
impact of profiling statements on middle end optimizations.