Reinstated value "MLton.debug", under the control of a command line constant.  Note that "MLton.debug" is _not_ affected by the -debug {false|true} flag.
Made lib/mlton/basic/{assert,trace}.sml enabled in the presence of
"MLton.debug" evaluating to true.

Added Makefile targets "traced" and "debugged".
The "traced" target produces a mlton compiler with the flags
"-const 'Exn.keepHistory true' -const 'MLton.debug true' -drop-pass 'deepFlatten'"
The -drop-pass 'deepFlatten' accomodates a performance problem, which
is probably related to the one seen by Martin Elsman when compiling
the MLKit: (http://mlton.org/pipermail/mlton/2005-April/026990.html)
The "debugged" target adds "-debug true" to the flaggs of the "traced"
target.

The "traced" target is provided because compiling with "-debug true"
is annoying because the .c file and many .s files are too big for the
assembly format; this yields many assembler warnings, which ultimately
makes the assembly processes IO bound.  Also, the addition of the
debugging information yields a 103MB executable, which strips to a
29MB executable.  It is probably worth investigating means by which we
can easily produce an executable that is linked with the debug
runtime, but does not itself have debug information.

Finally, attempted to regularize the string constants associated with
Trace.*, Assert.*, and Error.bug.  For the most part, they follow a
format like the following:
  "Functor/Structure.function: msg"