Fix for "shrinker didn't simplify right" error, which can be triggered by appling closure conversion to an insufficiently optimized input SXML program.
For example, if the input program looks like this:

fun F_0 (): {raises = None, returns = None} = L_8906 ()
L_8906 ()
  ...
  x_13339: unit = ()
  x_13338: unit ref = Ref_ref(unit) (x_13339)
  exnRef_35: unit ref = x_13338
  x_13337: exn = Bind_0 (exnRef_35)
  x_13336: exn = x_13337
  ...
  x_12660: unit = (...,
                   x_13339,
                   x_13338,
                   exnRef_35,
                   x_13337,
                   x_13336,
                   ...)
  return x_12660

then the SSA shrinker returns this:

fun F_0 (): {raises = None, returns = None} = L_8906 ()
L_8906 ()
  ...
  x_13339: unit = ()
  x_13338: unit ref = Ref_ref(unit) (x_13339)
  x_13337: exn = Bind_0 (x_13338)
  ...
  x_12660: unit = (...,
                   x_13339,
                   x_13338,
                   x_13338,
                   x_13337,
                   x_13337,
                   ...)
  return x_12660

It is incorrect to drop any variables that appeared in the return
tuple, as they might be referenced by other SSA functions.

The solution is to augment the statements returned by the shrinker
with additional bindings for variables eliminated.  That is, in the
example above, the collection of global statements would be:

  ...
  x_13339: unit = ()
  x_13338: unit ref = Ref_ref(unit) (x_13339)
  x_13337: exn = Bind_0 (x_13338)
  ...
  ...
  exnRef_35: unit ref = x_13338
  x_13336: exn = x_13337
  ...

These extraneous assignments will be killed off by the removeUnused
pass if they are truly unused in the rest of the program.