Added expert option  -opt-passes {default|minimal}, which allows simultaneous setting of XML, SXML, SSA and SSA2 optimization passes.  "default" is our current set of optimizations. "minimal" skips almost every optimization pass.  The exceptions are:  * deadCode  * sxmlShrink2 which runs just before implementExceptions;    The shrink pass is necessary to beta-reduce the "Exn_setInitExtra"    primitive so that it is applied to a ground term, as required by the    implementExceptions pass.  * constantPropagation, which implements Array_array0  * rssaShrink1,rssaShrink2;    the rssa simplify pass hasn't been refactored to match the other IL    simplify passes, so it would have been cumbersome.  * the x86 native codegen optimization passes
Fixed a few latent bugs in the backend and bytecode codegen that were
revealed by insufficient optimization of the input program.

The only compile errors I am seeing now are occasional unbound symbols,
which arise from _import or _prim expressions that are not simplified
away, but are not implemented on a particular os, arch, or codegen.  The
only regression failure I've seen is in real.sml with the native
codegen.  The comparison
  *+ (r1, r2, r3) == r1 * r2 + r3
fails, which makes sense, as *+ is implemented as a primitive, while in
the absence of inlining, * and + will each be implemented as ML function
calls to the corresponding primitive (requiring the arguments and
results to be saved to memory at 64-bits, rather than the 80-bits at
which the *+ primitive will compute).

This should serve as a starting point for anyone interested in designing
a good set of optimization passes that reduce compile times.  One
certainly doesn't want to drop all optimization passes -- the resulting
larger program will only slow down the necessary passes, and the runtime
performance will be unacceptable -- but it does serve as a good lower
abound and stress test of the compiler.