MAIL Improved FFI.
Incorporated Wesley's patch for improved FFI.  After the discussion on
the list regarding C pointers, I went ahead and eliminated _address
infavor of _symbol, which provides the address, getter, and setter.
So, the FFI system now looks like:

_import "symbol" [cdecl | stdcall] : cfTy;  ==>  cfTy
_import * [cdecl | stdcall] : ptrTy -> cfTy;  ==>  ptrTy -> cfTy
_export "symbol" [cdecl | stdcall] : cfTy;  ==> cfTy -> unit
_symbol "symbol" [define] : ptrTy, cbTy;  ==> ptrTy * (unit -> cbTy) * (cbTy -> unit)
_symbol * : ptrTy, cbTy;  ==> (ptrTy -> cbTy) * (ptrTy * cbTy -> unit)

I'm tempted to revise _import * to

_import * [cdecl | stdcall] : ptrTy, cfTy;  ==>  ptrTy -> cfTy

but I don't think it is crucial.