Added
  val MLton.GC.setRusage: bool -> unit

This sets a flag in the GC state that controls whether rusage
information is gathered for each collection.  The default is FALSE,
which is different than earlier behavior, but probably makes sense
because the getrusage calls at each GC can be costly, because
MLton.Rusage.rusage is not so used, and because it's easy to enable
the old behavior by calling MLton.GC.setRusage false.

This commit fixes a long-standing bug in the runtime in which
getrusage was called at every GC.  This was caused by our attempts to
use gcc's weak attribute to cleverly decide whether or not to measure
GC rusage, and a mistake in CPP #defines (HAS_WEAK vs SUPPORTS_WEAK).
In any case, we no longer use weak attributes, which weren't very
portable and didn't work as well as the above (which will do what the
user wants whether or not MLton.Rusage.rusage happens to appear in the
executable).



SVN r3995