Fixed Subscript bug in signature matching.
The bug was tickled by the following program, which caused an
unhandled exception to be raised.

  signature X =
     sig
        type x = unit
     end
  
  structure X :> X =
     struct
        type 'a x = unit
     end

The problem was in the isPlausible function, introduced back in
revision 3744.  It was checking schemes too early, under the
assumption that the type arities were equal, rather than waiting until
after the check that verified that they were (which in the above case
would fail).  The fix was to delay the checkSchemes call until after
isPlausible succeeds.



SVN r4086