Eliminated the "unused" field of GC_weak.
Instead, the representation for weaks is chosen by the backend to
respect the alignment (and, in the future, the objptr model).  Uses of
GC_weak in the runtime are handled by computing an offset from the
start of the object, again, respecting the alignment.

The upshot is that weak pointers require 4bytes less space with 
-align 4.

We could, and probably should, do something similar with GC_thread.



SVN r4204