Refactoring.
 * Ensure that primitives and primitive FFI imports make no assumption
   about default sizes.
 * Ensure that bitsize related characteristics are expressed in Int32/Word32;
   this includes shift arguments.
 * Major reworking of IntInf code to be parametric with respect to
   objptr size and mplimb size.

This is using a "poor-man's" functor approach via the config/* and
map/* files.  The Makefile includes a type-check target that
type-checks the basis library under a variety of different
representation choices.  This ensures that although we use transparent
structure assignment (to facilitate rebinding of structues as more
operations have been defined), we use the appropriate coercions where
necessary.



SVN r4348