MAIL preliminary support for compiler specific annotations
Added very simple support for compiler specific annotations.  If an
annotation contains ":", then the text preceding the ":" is meant to
denote a compiler.  For MLton, if the text preceding the ":" is equal
to "mlton", then the remaining annotation is scanned as a normal
annotation.  If the text preceding the ":" is not-equal to "mlton",
then the annotation is ignored, and no warning is issued.



SVN r4349