Sped up the implementation of Real.{==,!=,isNan}.
Real.== no longer calls Real.class to ensure basis spec compliance.
Instead, it assumes that the primitive == does the right thing.

Real.isNan is now implemented with

  fun isNan r = r != r

This was about 4 times faster on my machine than the old version that
used Real.class.

Changed the primitive simplifier so that it doesn't simplify
Real_equal(x, x) to true, since this is incorrect for nans.



SVN r4385