Refactored PackWord.
Implemented PackWord structures using C-functions to sub and update.
This follows the implementation of PackReal.  In the past, we've used
primitives for PackWord32.  We will likely use primitives again in the
future, but its easier to get these architecture dependent primitives
out of the way for the time being.



SVN r4418