Ported to PowerPC/AIX.
Fixed a bug in the runtime for the cases where nonblocking IO with
sockets was implemented using MSG_DONTWAIT.  This flag does not exist
on AIX, Cygwin, HPUX, and MinGW and was previously just ignored.  Now
the runtime simulates the flag for these platforms (except MinGW, yet,
where it's still ignored).

Some cosmetics.


SVN r4536