Tweaked the assert macro to call a dummy "asok" function in the event that the assertion is true.  This avoids spurious warnings like the following:
  gc/mark-compact.c:18: warning: statement with no effect
  gc/mark-compact.c:30: warning: statement with no effect

These warnings come when compiling with -pedantic -Wall, and appear to
be due to gcc simplifying the assert test (p) in

  (p) ? (void)0 : asfail(__FILE__, __LINE__, #p)

to TRUE at compile time and replacing expression with (void)0.  Now,
why this should cause a warning is beyond me, since (void)0 by itself
as source code doesn't.

Anyways, introducing the call to "asok" instead of (void)0 suppresses
the warning.  And the performance shouldn't matter since this is an
assert and is only on when debugging.



SVN r4542