In r4571:
  Calling translateHeap with s->heap.oldGenSize is incorrect, because
  when heapRemap fails and heapCreate succeeds, s->heap is updated with
  an newly initialized heap with oldGenSize == 0.  Instead, call
  translateHeap with size, which has the original oldGenSize.

The commit message failed to note that it additionally set the newly
initialized heap's oldGenSize to size.  Leaving the newly initialized
heap with oldGenSize == 0 is incorrect, at it makes the heap appear to
have no data.  Once the newly initialized heap's oldGenSize is set
correctly, then it is acceptable to call translateHeap with
s->heap.oldGenSize.



SVN r4577