This commit gets MinGW to completely rebuild.
The first change switches world & profiling code to use stdio:
  1. stdio is ANSI C and thus more portable
  2. the raw io used previously did not
   a. deal with binary/text mode under windows (=> corrupt world files)
   b. restart large reads/writes when they only partially completed
   c. align writes on sector boundaries; Linus threatened to break this!

The second change is how the SML code saves the world:
Now the filename is passed to C, and a failure does not abort the program.
Instead, we check return codes and propogate the error code back to an SML
exception, raised with the correct error status.

To effect this change, I had to change the save primitive's prototype.
Also, I had to output nicer C code in the codegen to keep error status.



SVN r4642