Fixed problem introduced at r4642:
  The second change is how the SML code saves the world:
  Now the filename is passed to C, and a failure does not abort the program.
  Instead, we check return codes and propogate the error code back to an SML
  exception, raised with the correct error status.

The problem is due to changing the primtive from
  val save = _prim "World_save": C.Fd.t -> unit
to
  val save = _prim "World_save": NullString8_t -> bool C_Error.t

It is not possible to have the type of save as 
NullString8.t -> bool C_Errno.t, because there are two different ways
to return from the call to save.  One way is the direct obvious way,
in the program instance that called save.  However, another way to
return is in the program instance that loads the world.  Making save
return a bool creates nasty bugs where the return code from the CCall
expects to see a bool result according to the C return convention, but
there isn't one when returning in the load world.

So, save's result status is accessible via:
  val getSaveStatus =
     _import "GC_getSaveWorldStatus" : GCState.t -> bool C_Errno.t



SVN r4654