Added the new build/lib/$crossTarget/include directory when using gcc to compile cross-built C files.
Fixed bug introduced in bin/regression in r4515.  The relevant comment
in the commit was:

> 4. Potentially modified the behaviour of bin/regression
> 	A comment previously stated:
>                 #   Must use $f.[0-9].[cS], not $f.*.[cS], because the
>                 #   latter will include other files, e.g. for finalize,
>                 #   it will also include finalize.2.
> 	My testing shows this to be false. Furthermore, the work-around:
>                 files="$f.[0-9].[cS]"
>                 if [ 0 -ne `ls $f.[0-9][0-9].[cS] 2>/dev/null | wc -l` ]; then 
>                         files="$files $f.[0-9][0-9].[cS]"
>                 fi
> 	will break on a filename with spaces or shell specials.

In fact, the problem does happen, but only when cross compiling the
regressions.  I went back to the old workaround, which as the above
notes, is wrong for filenames with spaces or specials.  I'd be happy
to see that fixed.



SVN r4668