Moved the jumpToSML/returnToC code from x86-main.h to x86-codegen.fun.
This code essentially establishes the interface by which C jumps into
SML code (both initially and for exported SML functions) and by which
SML returns to C (for exported SML functions).  The key property is
that all C activation frames must be accounted for on the C stack.
Each time C jumps to SML via jumpToSML, the current C stack pointer
is saved for C calls from SML.  Additionally, we ensure that the saved
C stack pointer is 16-byte aligned.

Previously, this logic was handled in x86-main.h via inline assembly.
This had a number of disadvantages:
 1. We couldn't be sure of what gcc was doing with %esp before the
inline assembly began.  In fact, we couldn't be sure that %esp
necessarily pointed to the top of the C stack.
 2. We needed extra logic from CPP and from MLton_main arguments to
determine whether or not %esp was reserved for sigstack handling and
whether or not extra "_" were required by assembly labels.

Moving the definition of this code to x86-codegen.fun solves these
problems:
 1. Since we control the entire execution, we know that %esp is
16-aligned on entry (modulo the return address pushed by the call
instruction) and can maintain this property easily.
 2. We can save temporaries in the C stack frame corresponding to the
call to MLton_jumpToC.
 3. We know whether or not to reserve %esp and whether or not labels
need an extra "_".

The downside is that writing out the assembly via the datatype
interface is much more verbose.



SVN r4718