Adam Goode's Makefile patches:
Replace all instances of plain old "=" with ":=", except in
packages/freebsd/Makefile since I have no way of testing it. This
should result in faster build times and less possibility for confusion
in the future. Note that bytecode/Makefile actually relied on the
semantics of lazy assignment, so I had to reorder some lines in the
file. I don't believe any other Makefile in the MLton tree was
sensitive to this issue.

bytecode/Makefile is missing the TARGET_OS variable. This could cause
the build to fail strangely when not building from the toplevel
Makefile (which exports TARGET_OS). runtime/Makefile correctly defines
this variable.



SVN r4863