Added new expert option: -prefer-abs-paths {false|true}
Setting the option to true is supposed to have the effect that source
files are referred to by their absolute paths.  This can simplify
(working with and implementation of) external tools that would
otherwise have to know the directory from which the compiler was
executed (in order to locate source files).


SVN r5076