MAIL added option -profile-val
Some very minor additions to the profiling infrastructure, mainly
addressing a comment raised by:
  http://mlton.org/pipermail/mlton-user/2007-January/000992.html

The '-profile-val true' option instructs the compiler to push and pop
profiling information for the evaluation of (expansive) val bindings.
With the program described in the message above, we now have the
following output (with "-const 'Exn.keepHistory true' -profile-val true"):

[fluet@shadow tmp]$ ./z
unhandled exception: Uninitialized
with history:
	r.<raise> z.sml 5.36
	r z.sml 4.9
	<val>:z z.sml 9.9
	<main>

So, we get the very useful information about which top-level binding
is responsible for the exception.

With "-profile-val true -profile-stack true", it should be possible to
distinguish expensive calls to the same function.

I'm not 100% happy with the syntax, so suggestions are welcome.

I also tweaked basis-library/mlton/call-stack.sml to only ignore one
element at the bottom of the call stack; this seems to have solved the
issue with "shallow" call stacks.



SVN r5257