Ensure that setInitExtra and setExtendExtra are initialized.
This is required for situation where the Basis Library raises (and
handles) exceptions during initialization.

(For example, raising and handling Overflow to compute Array.maxLen
from SeqIndex.maxInt using SeqIndex.toInt.  Thraises Overflow if
SeqIndex == Int64 and Int == Int32, which indicates that we should use
Int.maxInt for Array.maxLen.  On the other hand, this does not raise
Overflow if SeqIndex == Int32 and Int == Int64, which indicates that
we should use SeqIndex.toInt (SeqIndex.maxInt) for Array.maxLen.)

Befor this commit, when such a program is compiled with "-const
'Exn.keepHistory true'", it would immediately terminate with
'MLton_bug "setExtendExtra unimplemented"'.  This corresponds to
trying to raise an exception before basis-library/mlton/exn.sml
executed, which would setExtendExtra.

Now, such a program is able to raise an exception during
initialization (without an associated call-stack).



SVN r5369