Going to 64-bit headers and array lengths on 64-bit executables.
See commit r5322 for more details on preliminary 64-bit executables.

This commit "fixes" the mark-compact collector.  The issue is that the
mark-compact collector threads object pointers through headers.

We might be able to get by with 32-bit headers, if every object was
required to have 32-bits of non-object-pointers after the header.  (It
would be o.k. if these bits were from object data.)  However, this
wouldn't suffice for an array of object pointers, since adding another
32-bits at the beginning of the array would break offsets; and adding
32-bits to every element (just to get 32-bits at the beginning of the
object) is too much of a space waste.

So, we need to go with 64-bit headers.  However, advanceToObjectData
assumes that one can distinguish between the beginning of a normal
object header and the beginning of an array object header by looking
for the 0 of the array counter word.  So, array counters need to be
64-bits, and since the compiler currently assumes that the array
counter and the array length fields of an array header are the same
size, array lengths need to be 64-bits.

Note that the default integer size is still Int32.int, so Array.maxLen
(and array indexes) are still only 32-bits (from user code).

Haven't tested very large arrays with -default-ty int64.



SVN r5370