Patch from Nicolas Bertolotti (PolySpace).
- bug-fix-times-mingw.patch

This patch is designed to avoid raising an exception during the
initialization of the basis library on MinGW when the code contains calls to
the "times" function which are not detected as dead code during compilation.
Then, if the "times" function is called, the exception will be raised from
the location of the call which is easier to understand from a user's point
of view.



SVN r5488