Native AMD64 codegen.
This commit changes the amd64-* target platforms to use the native
amd64 codegen.

All regressions compile and pass (with -debug true); a self-compile
bootstraps without issue.

The amd64 codegen is basically a copy/paste/edit of the x86 codegen.
There is probably some commonality that could be factored, but there
are lots of minor difference scattered everywhere.  And, heck, MLton
would defunctorize and monomorphise everything anyways, so we aren't
likely to save compile time or code size.

The most significant difference between the x86 codegen and the amd64
codegen is that on amd64, we use the SSE2 instructions and XMM
registers for floating-point operations, entirely ignoring the x87
floating point stack.  This supports the basic operations
(+,-,*,/,==,?=,<,<=), but all MATH functions go through libmath C
functions (gcc does the same thing).  The nice bonus is that this
gives IEEE compilant floating-point operations; i.e., Real64 ops
compute at 64-bits of precision and Real32 ops compute at 32-bits of
precision.

The other big difference is more registers, all with 64/32/16/8 bit
variants.  The 8-bit high registers of RAX, RBX, RCX, and RDX are
ignored.  This very minorly complicates 8-bit multiplication and
division, which use AH rather than DL (in contrast to the 16/32/64-bit
multiplication and division, which use RDX).  When an 8-bit mult or
div is emitted to the .S file, AH is copied to/from DL in the
appropriate manner.

The -codegen option has changed from
  -codegen {native|c|bytecode}
to
  -codegen {x86|amd64|c|bytecode}
in order to distinguish the two native codegens.



SVN r5566