Initial commit of bg-build-mode for ``handsfree'' background builds.
If you don't want to manually start/stop/restart builds after modifying
files, then this mode is for you.  Just save the file, a compile is
started (silently!), you can continue working without even thinking about
starting a build, and if there are errors, you are notified (with a
message) and you can then jump to errors.

This mode is not specific to MLton per se, but is particularly useful for
working with MLton due to the longer compile times.  By the time you start
wondering about possible errors, the build is already on the way.


SVN r5681