Fixed a *very* long-standing bug in the monomorphisation pass, reported by Vesa Karvonen.  (AFAICT, the bug has been present since the original mlton-19990319 release.)
Unlike value variables, type variables are not uniquified by the
front-end passes.  Hence, it is entirely possible for a type variable
in the XML IL (that is a symbol with an associated property list) to
be used as a type variable for multiple polymorphic variables.
Indeed, since explicit type variables in the source derive their
identity from the corresponding source type variable (as a textual
name), there are often many uses of the "'a" type variable throughout
the XML IL program, though with different scopes.

The monomorphisation pass works by setting (via a property) a
monomorphic instantiation for a type variable, which is meant to
persist for the scope of the type variable.  Unfortunately, by
monomorphising a sub-expression, we may end up setting a different
monomorphic instantiation for the same type variable (but
corresponding to a different scope).  There are occasions where the
monomorphisation pass returns from monomorphising a sub-expression to
monomorphise a type.  If the monomorphisation of the sub-expression
set a type variable used by the type, then the type is monomorphised
under the wrong monomorphic instantiation.

It turns out that monomorphising a type never results in setting a
type variable's monomorphic instantation.  Hence, for a PolyVal, it
suffices to monomorphise the type before monomorphising the
expression.  (This change suffices for Vesa's code.)

For a Fun (i.e., a recursive collection of polymorphic functions),
however, there are multiple calls to monomorphise a lambda, each of
which may reset a type variable's monomorphic instantiation.  Hence,
the proper monomorphic instantiation must be esablished for each
lambda.  (This wasn't triggered by Vesa's code.)



SVN r5731