Use generalized gdtoa functions to perform conversions at specified rounding modes. Only Real.{fmt StringCvt.EXACT,fromDecimal,toDecimal} are specified to operate under TO_NEAREST semantics. Real.{fmt StringCVT.{FIX,GEN,SCI},fromString,scan,toString} should all operate under the semantics of the current rounding mode.
This means that while
  Real.fromDecimal (IEEEReal.fromString (Real.fmt StringCvt.EXACT r))
and
  Real.fromDecimal (Real.toDecimal r)
and
  Real.fromDecimal (IEEEReal.fromString (IEEReal.toString (Real.toDecimal r)))
are all guaranteed to return r (under any ambient rounding mode), it
is not the case that
  Real.fromString (Real.fmt StringCvt.EXACT r)
is guaranteed to return r; the exact decimal approx (as a string) need
only convert to r under the TO_NEAREST rounding mode.

However,
  Real.fromString (Real.fmt (StringCvt.{FIX,GEN,SCI} (SOME 1000)) r)
should return r (under any ambient rounding mode); that is, a
real->string conversion (with sufficient digits of precision) will
yield a string that converts back to the same real at the same
rounding mode.  (The PLDI'90 papers term this 'correct rounding', and
the gdtoa library should provide it.)



SVN r5827