Tweaks to grab-wiki and make-pdf-guide scripts.
For grab-wiki, save the date of the grab in MLTONWIKIVERSION and stamp
with the string "MLTONWIKIVERSION" in the upper left of pages and for
ViewSVN references, replace reference to "trunk" with
"on-MLTONWIKIVERSION-release".

For make-pdf-guide, replace the string "MLTONWIKIVERSION" with the
date from the file MLTONWIKIVERSION.

This should have the effect that only the pages that actually change
between runs of grab-wiki will need to be committed to SVN (rather
than all files, since all files were being stamped with the current
date).



SVN r5910